backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "edge-tts>=7.2.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "google-genai>=1.38.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pycharacterai>=2.2.50",
]
